+++ Commons/src/Commons.Core/NumericalMethods/DifferentialEquation/DifferentialEquationMethod.cs
@@ -5,12 +5,20 @@ namespace MudRunner.Suspension.Core.NumericalMethods.DifferentialEquation
     /// <summary>
     /// It is responsible to execute numerical method to solve Differential Equation.
     /// </summary>
-    /// <typeparam name="TInput"></typeparam>
-    public abstract class DifferentialEquationMethod<TInput> : IDifferentialEquationMethod<TInput>
-        where TInput : NumericalMethodInput
+    public abstract class DifferentialEquationMethod : IDifferentialEquationMethod
     {
+        /// <summary>
+        /// Dimensionless.
+        /// </summary>
+        protected abstract double Gama { get; }
+
+        /// <summary>
+        /// Dimensionless.
+        /// </summary>
+        protected abstract double Beta { get; }
+
         /// <inheritdoc/>
-        public virtual NumericalMethodResult CalculateInitialResult(TInput input)
+        public virtual NumericalMethodResult CalculateInitialResult(NumericalMethodInput input)
         {
             return new NumericalMethodResult
             {
@@ -22,6 +30,6 @@ namespace MudRunner.Suspension.Core.NumericalMethods.DifferentialEquation
         }
 
         /// <inheritdoc/>
-        public abstract Task<NumericalMethodResult> CalculateResultAsync(TInput input, NumericalMethodResult previousResult, double time);
+        public abstract Task<NumericalMethodResult> CalculateResultAsync(NumericalMethodInput input, NumericalMethodResult previousResult, double time);
     }
 }