+++ Commons/src/Commons.Core/NumericalMethods/DifferentialEquation/NewmarkBeta/INewmarkBetaMethod.cs
@@ -1,9 +1,7 @@
-﻿using MudRunner.Suspension.Core.Models.NumericalMethod.NewmarkBeta;
-
-namespace MudRunner.Suspension.Core.NumericalMethods.DifferentialEquation.NewmarkBeta
+﻿namespace MudRunner.Suspension.Core.NumericalMethods.DifferentialEquation.NewmarkBeta
 {
     /// <summary>
     /// It is responsible to execute the Newmark-Beta numerical method to solve Differential Equation.
     /// </summary>
-    public interface INewmarkBetaMethod : IDifferentialEquationMethod<NewmarkBetaMethodInput> { }
+    public interface INewmarkBetaMethod : IDifferentialEquationMethod { }
 }